Updated play screens, added new craft screens
The wool icon on the play screens is now 60px tall, the text is to
display inside of each ball of yarn.

Each tab of the craft screen, e.g., craft-hats-dev.png, displays inside
the orange area in craft-dev.

The total coins are displayed in the gold frame in footer area.