Added store background, updated craft background
Base layers are store.png and craft.png

Each has a close (button) button in the upper-left corner.
The total counts of wool in the upper-center.
The total counts of eweros in the bottom-center.

The store screen has a sign, "Craft Catalog" to switch to the craft
screen.

The craft screen has a sign, "General Store" to switch to the store
screen.