Added new craft screens and hats images
Each tabbed page is named craft-{items}.png, e.g., craft-hats.png. Each
of these overlay craft.png the top-left corner is 80px from the top.

Each tab has a dev view, e.g., craft-hats-dev, with color-coordinatied
areas. Below are descriptions of what appears in each area.

Items are in two sizes, i.e., -small and -large suffix, e.g.,
hat-white-red-small.png and hat-white-red-large.png. -small images
appear over the swatch of their color combination, each is 60px x 60px.

Items that can't be crafted are {item}-disabled-small and
{item}-disabled-large, e.g., hat-disabled-small.png and
hat-disabled-large.png.

When an item has been crafted once, craft-star.png will overlay the
green area next to the item.

The craft button will appear In the blue area with pink rectangle on
the right. When an item can be crafted button-craft.png will display.
When an item can't be crafted, button-craft-disabled.png will display.

Above the craft button, aligned with the top-left corner of the orange
rectangle is where the large item will display, e.g.,
hat-white-red-large.png. The large item will display when its small
item is clicked. If the item can't be crafted, its image will be
replaced with {item}-disabled-large.png, e.g., hat-disabled-large.png.

In the upper-right corner of the orange rectangle, inside the pink
rectangle, is where an items craft value appears. It is represented in
Eweros (coins). If an item's value represented as dollars was $5.50 it
will be represented in Eweros as 550 (no $ or .).